Addition of detachment flag to parameter set
To kill a detached dynamodb, use the stopChild command with the child.
Highly recommended to use the async await functionality from ES8+ as it
saves a lengthy promise chains.
Fixed the verbose flag, to default to false
Fixed a type on the child.close that stated start instead of close
Added -Xrs flag to be able to kill the child without running into exit
code 143.